refactor(backend): split learning store row mappers
